Added check for ownership while editing or deleting a user

+++ controllers/user.js
@@ -16,28 +16,34 @@ export const getParticularUser = async (req, res) => {
 
 export const editUserDetails = async (req, res) => {
   try {
-    const { username, ...data } = req.body;
+    if (req.params.id != res.locals.user.id)
+      return res.status(401).json({ success: false, error: "Unauthorized" });
 
-    const editedUserDetails = await User.findByIdAndUpdate(
+    const editDetails = { status: req.body.status, name: req.body.name };
+
+    const editedUser = await User.findByIdAndUpdate(
       req.params.id,
-      { $set: data },
+      { $set: editDetails },
       { returnDocument: "after" }
     );
 
-    res.status(200).json(editedUserDetails);
+    res.status(200).json({ success: true, editedUser });
   } catch (err) {
     console.log(err.message, " on Route ", "'PATCH /user/:id'");
-    res.status(400).json({ success: false, error: err.message });
+    res.status(404).json({ success: false, error: "Not Found" });
   }
 };
 
 export const deleteUser = async (req, res) => {
   try {
+    if (req.params.id != res.locals.user.id)
+      return res.status(401).json({ success: false, error: "Unauthorized" });
+
     const deletedUser = await User.findByIdAndDelete(req.params.id);
 
     if (!deletedUser) throw Error();
 
-    res.status(200).json(deletedUser);
+    res.status(200).json({ success: true, deletedUser });
   } catch (err) {
     console.log(err.message, " on Route ", "'DELETE /user/:id'");
     res.status(404).json({ success: false, error: "Not Found" });
